Rename to DeprecationHelperRemoveRector and support full versions instead of only majors
