+++ src/Rector/Deprecation/DeprecationHelperRemoveRector.php
@@ -3,19 +3,20 @@
 namespace DrupalRector\Rector\Deprecation;
 
 use DrupalRector\Rector\AbstractDrupalCoreRector;
-use DrupalRector\Rector\ValueObject\RemoveDeprecationHelperConfiguration;
+use DrupalRector\Rector\ValueObject\DeprecationHelperRemoveConfiguration;
 use PhpParser\Node;
 use PhpParser\NodeDumper;
 use Rector\Core\Contract\Rector\ConfigurableRectorInterface;
 use Rector\Core\Rector\AbstractRector;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
+use Symplify\RuleDocGenerator\ValueObject\CodeSample\ConfiguredCodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
-final class RemoveDeprecationHelperRector extends AbstractRector implements ConfigurableRectorInterface
+final class DeprecationHelperRemoveRector extends AbstractRector implements ConfigurableRectorInterface
 {
 
     /**
-     * @var array|RemoveDeprecationHelperConfiguration[]
+     * @var array|DeprecationHelperRemoveConfiguration[]
      */
     private array $configuration;
 
@@ -24,19 +25,26 @@ final class RemoveDeprecationHelperRector extends AbstractRector implements Conf
      */
     public function getRuleDefinition(): RuleDefinition
     {
-        return new RuleDefinition('Fixes deprecated user_password() calls',[
-            new CodeSample(
+        return new RuleDefinition('Remove DeprecationHelper calls for versions before configured minimum requirement', [
+            new ConfiguredCodeSample(
                 <<<'CODE_BEFORE'
 $settings = [];
 $filename = 'simple_filename.yaml';
-drupal_rewrite_settings($settings, $filename);
+DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', fn() => old_function(), fn() => new_function());
+DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '10.1.0', fn() => drupal_rewrite_settings($settings, $filename), fn() => SettingsEditor::rewrite($filename, $settings));
+DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '11.1.0', fn() => old_function(), fn() => new_function());
 CODE_BEFORE
                 ,
                 <<<'CODE_AFTER'
 $settings = [];
 $filename = 'simple_filename.yaml';
-DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '10.1.0', fn() => drupal_rewrite_settings($settings, $filename), fn() => SettingsEditor::rewrite($filename, $settings));
+drupal_rewrite_settings($settings, $filename);
+DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '11.1.0', fn() => old_function(), fn() => new_function());
 CODE_AFTER
+                ,
+                [
+                    new DeprecationHelperRemoveConfiguration('10.1.0')
+                ]
             )
         ]);
     }
@@ -44,10 +52,10 @@ CODE_AFTER
 
     public function configure(array $configuration): void {
         foreach ($configuration as $value) {
-            if (!($value instanceof RemoveDeprecationHelperConfiguration)) {
+            if (!($value instanceof DeprecationHelperRemoveConfiguration)) {
                 throw new \InvalidArgumentException(sprintf(
                     'Each configuration item must be an instance of "%s"',
-                    RemoveDeprecationHelperConfiguration::class
+                    DeprecationHelperRemoveConfiguration::class
                 ));
             }
         }
@@ -83,8 +91,7 @@ CODE_AFTER
             assert($args[1]->value instanceof Node\Scalar\String_);
 
             $introducedVersion = (string) $args[1]->value->value;
-            $introducedVersionParts = explode('.', $introducedVersion);
-            if ((int) array_shift($introducedVersionParts) + 1 !== $configuration->getMajorVersion()) {
+            if (version_compare($introducedVersion, $configuration->getMinimumRequirement(), '>=')) {
                 continue;
             }
