+++ tests/src/Rector/Deprecation/DeprecationHelperRemoveRector/fixture/user_password.php.inc
