+++ tests/src/Rector/Deprecation/DeprecationHelperRemoveRector/fixture/user_password_keep.php.inc
@@ -14,10 +14,10 @@ function simple_example()
 
 function simple_example()
 {
-    $password = \Drupal\Component\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '10.1.0', fn() => user_password(), fn() => \Drupal::service('password_generator')->generate());
-    $other_password = \Drupal\Component\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '10.1.0', fn() => user_password(8), fn() => \Drupal::service('password_generator')->generate(8));
+    $password = \Drupal::service('password_generator')->generate();
+    $other_password = \Drupal::service('password_generator')->generate(8);
     $password_length = 12;
-    $last_password = \Drupal\Component\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '10.1.0', fn() => user_password($password_length), fn() => \Drupal::service('password_generator')->generate($password_length));
+    $last_password = \Drupal::service('password_generator')->generate($password_length);
 }
 
 ?>
